refactor: extract project file actions into a hook
